warnings + project reorg + win7 taskbar
* removed a few warnings
* reworked the project files for ease of use.
* added in an explicit win 7 app id
  allows creation of other programs to manipulate the tasks/jumplists

+++ WINDOWS/WINNET.C
@@ -1681,7 +1681,7 @@ int net_service_lookup(char *service)
 
 char *get_hostname(void)
 {
-    int len = 128;
+    size_t len = 128;
     char *hostname = NULL;
     do {
 	len *= 2;
